requirements.txt
Django==4.2.6
djangorestframework==3.15.0
psycopg2-binary==2.9.8
django-cors-headers==3.14.0
Pillow==10.0.1
